correcion de errores, seguridad

+++ app/Http/Livewire/Admin/Estudiantes/Esteliminar.php
@@ -2,6 +2,8 @@
 
 namespace App\Http\Livewire\Admin\Estudiantes;
 
+use App\Models\Ficha;
+use App\Models\Seguimiento;
 use App\Models\User;
 use Illuminate\Support\Facades\Storage;
 use Livewire\Component;
@@ -25,7 +27,26 @@ class Esteliminar extends Component
     // ! FALTA DESVINCULAR TUTORADOS EN CASO TENGA TUTORADOS A CARGO
     public function eliminarEst()
     {
-        Storage::disk('public')->delete($this->est->foto);
+        if ($this->est->foto) {
+            // eliminar foto
+            Storage::disk('public')->delete($this->est->foto);
+        }
+
+        // eliminar ficha
+        $ficha = Ficha::where('user', $this->est->id)->first();
+        if ($ficha) {
+            Ficha::destroy($ficha->id_ficha);
+        }
+
+        // eliminar los seguimientos
+        $seguimientos = Seguimiento::where('user', $this->est->id)->get();
+        if ($seguimientos) {
+            foreach ($seguimientos as $seg) {
+                Seguimiento::destroy($seg->id_seg);
+            }
+        }
+
+        // eliminar user
         User::destroy($this->est->id);
 
         session()->flash('cerrarModal');
